feat: separate cli and runtime
