+++ cli/example.js
